Consolidate forms TS types
