+++ projects/packages/forms/src/dashboard/integrations/creative-mail-card.tsx
@@ -2,7 +2,7 @@ import { Button } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import IntegrationCard from '../../blocks/contact-form/components/jetpack-integrations-modal/integration-card';
 import CreativeMailIcon from '../../icons/creative-mail';
-import type { IntegrationCardProps } from './types';
+import { IntegrationCardProps } from '../../types/types';
 
 const CreativeMailDashboardCard = ( {
 	isExpanded,
