--- projects/packages/forms/src/dashboard/types.ts
@@ -1,7 +1,0 @@
-export type Pattern = {
-	image: string;
-	title: string;
-	recommended?: boolean;
-	code: string;
-	description: string;
-};
