+++ projects/packages/forms/src/types/types.ts
@@ -1,5 +1,15 @@
-// Shared types for dashboard integrations
+// Consolidated types for the forms package
 
+// Type definition from dashboard/types.ts
+export type Pattern = {
+	image: string;
+	title: string;
+	recommended?: boolean;
+	code: string;
+	description: string;
+};
+
+// Shared types for dashboard integrations from dashboard/integrations/types.ts
 export interface Integration {
 	type: 'plugin' | 'service';
 	slug: string;
